feat(packages/sui-test): Move MockFetcherManager
